vulkan: Add support for timeline semaphores
Unlike binary semaphores, timeline semaphores represent a (64bit
unsigned) value and the binary decision about whether they have signaled
or not is decided by comparing their value to a predefined value.
So this code adds the ability to pass that value.

In fact, it now requires this value, but that is not a problem because
Vulkan ignores the value for binary semaphores, so we can just pass
0 everywhere a binary sempahore is used.

There is no use of the code in those commit, but timeline semaphores
are used for explicit sync with dmabufs and D3D12 uses them exclusively.