gdkglcontext-win32-wgl.c: Apply scale for glAddSwapHintRectWIN()
We still need to pass the scaled positions and size to
glAddSwapHintRectWIN(), as we do when we call Windows API functions to
position or resize surfaces.

This reverts the changes on gdkwin32glcontext-win32-wgl.c in MR !8244.

Fixes items not rendering or partially rendering when the GL renderer is
being used.